refactor data model packages
