+++ modules/api/src/main/java/org/eclipse/xpanse/api/exceptions/CommonExceptionHandler.java
@@ -12,9 +12,9 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
-import org.eclipse.xpanse.api.response.Response;
-import org.eclipse.xpanse.api.response.ResultType;
 import org.eclipse.xpanse.modules.models.common.exceptions.ResponseInvalidException;
+import org.eclipse.xpanse.modules.models.response.Response;
+import org.eclipse.xpanse.modules.models.response.ResultType;
 import org.eclipse.xpanse.modules.models.service.register.exceptions.TerraformScriptFormatInvalidException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.converter.HttpMessageConversionException;
