+++ modules/deployment/pom.xml
@@ -60,7 +60,7 @@
         <dependency>
             <groupId>org.eclipse.xpanse.modules</groupId>
             <artifactId>orchestrator</artifactId>
-            <version>1.0.0-SNAPSHOT</version>
+            <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
     </dependencies>
