+++ plugins/huaweicloud/pom.xml
@@ -66,7 +66,7 @@
         <dependency>
             <groupId>org.eclipse.xpanse.modules</groupId>
             <artifactId>deployment</artifactId>
-            <version>1.0.0-SNAPSHOT</version>
+            <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
