+++ runtime/src/test/java/org/eclipse/xpanse/runtime/OrchestratorApiTest.java
@@ -16,10 +16,10 @@ import java.util.Map;
 import java.util.UUID;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.xpanse.api.OrchestratorApi;
-import org.eclipse.xpanse.api.response.Response;
 import org.eclipse.xpanse.modules.database.register.RegisterServiceEntity;
 import org.eclipse.xpanse.modules.models.admin.SystemStatus;
 import org.eclipse.xpanse.modules.models.admin.enums.HealthStatus;
+import org.eclipse.xpanse.modules.models.response.Response;
 import org.eclipse.xpanse.modules.models.service.common.enums.Category;
 import org.eclipse.xpanse.modules.models.service.common.enums.Csp;
 import org.eclipse.xpanse.modules.models.service.deploy.CreateRequest;
